requirements.txt
numpy==1.19.4
matplotlib==3.3.3
opencv_python==4.4.0.46
torchvision==0.8.2
torch==1.7.1
Pillow==8.0.1
